requirements.txt
pymupdf
sqlite-vec
httpx
requests
python-dotenv
